Add model validation and more action results after exist check and validation errors

+++ BoardGamesApi/Controllers/GamesController.cs
@@ -17,6 +17,13 @@ namespace BoardGamesApi.Controllers
         [HttpDelete("{id}")]
         public IActionResult Delete(string id)
         {
+            if (string.IsNullOrWhiteSpace(id))
+                return NotFound();
+
+            var game = _gamesRepository.GetById(id);
+            if (game == null)
+                return NotFound();
+
             _gamesRepository.Delete(id);
 
             return Ok();
@@ -30,9 +37,26 @@ namespace BoardGamesApi.Controllers
             return Ok(games);
         }
 
-        [HttpPost]
-        public IActionResult Post(GameInput model)
+        [HttpGet("{id}")]
+        public IActionResult GetById(string id)
         {
+            if (string.IsNullOrWhiteSpace(id))
+                return NotFound();
+
+            var game = _gamesRepository.GetById(id);
+
+            if (game == null)
+                return NotFound();
+
+            return Ok(game);
+        }
+
+        [HttpPost]
+        public IActionResult Post([FromBody] GameInput model)
+        {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var game = new Game();
             model.MapToGame(game);
 
@@ -44,9 +68,15 @@ namespace BoardGamesApi.Controllers
         }
 
         [HttpPut("{id}")]
-        public IActionResult Put(string id, GameInput model)
+        public IActionResult Put(string id, [FromBody] GameInput model)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var game = _gamesRepository.GetById(id);
+
+            if (game == null)
+                return NotFound();
 
             model.MapToGame(game);
